requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
uwsgi>=2.0.19,<2.1
django-cors-headers>=4.2.0,<4.2.1
django-redis>=5.3.0,<5.3.1
channels>=4.0.0,<4.0.1
channels-redis>=4.1.0,<4.1.1
daphne>=4.0.0,<4.0.1
ytmusicapi
yt-dlp